Add files via upload
Updated the minimum time interval for the BLE connection to 10ms. This is to make sure that we are able to have continuous connection between the device and the collector. Otherwise, we were seeing the huge number of packet drop or one of the device taking up all the connection time with the collector.